[Test] Character 테스트 코드 작성
